compilable checkpoint for changes to support Ruleset-level filters.
Changes include introducing Response interface, and splitting the filtering logic into a separate FilteredResponse implementation (the old stuff got moved to RawResponse).  This split feels a bit unnecessary, except for the fact that ResponseRule will hold onto a stable instance of the FilteredResponse, but the RulesetRule will need to build the FilteredResponse with each request.  Therefore, splitting out the filtering logic made sense.